+++ changelog.md
@@ -24,6 +24,7 @@
 - missing events on app windows hook.
 - wasted space reserved for tiling window manager when dock/toolbar are disabled on monitor.
 - fix missing variables values on default percentages for theme variables.
+- #1290 dock and toolbar showing on other monitors when mouse is at edge of screen.
 
 ## [2.4.9]